Split out into OldSchoolScreen class, hopefully will be able to use 'profiles' to emulate various old scholl machines, not just Commodore 64 (liek ZX Spectrum, Microbee, etc...).
