+++ docs/books/adventure-games/fantasy-games/screen_profiles.py
@@ -1,0 +1,544 @@
+COMMODORE_64 = {
+    # Commodore64 'full screen' is 403 pixels wide x 284 pixels high , including the border area
+    'full_screen_size': (403, 284),
+    # Commodore64 'addressable' screen area - 320 pixels wide x 200 pixels high
+    'screen_size': (320, 200),
+    # Commodore64 text characters are 8 pixels wide x 8 pixels high
+    'character_size': (8, 8),
+    # Commodore 64 colors - see the very interesting article here:
+    #    http://unusedino.de/ec64/technical/misc/vic656x/colors/
+    'colors': [
+        '#000000',  # black
+        '#FFFFFF',  # white
+        '#68372B',  # red
+        '#70A4B2',  # cyan
+        '#6F3D86',  # purple
+        '#588D43',  # green
+        '#352879',  # blue
+        '#B8C76F',  # yellow
+        '#6F4F25',  # orange
+        '#433900',  # brown
+        '#9A6759',  # light red
+        '#444444',  # dark grey
+        '#6C6C6C',  # grey
+        '#9AD284',  # light green
+        '#6C5EB5',  # light blue
+        '#959595',  # light grey
+    ],
+    # Commodore 64 character/font as bytes -
+    # TODO these don't all seem to correspond to the correct ASCII/PETSCII character locations
+    'font': [
+        [60, 102, 118, 118, 6, 70, 60, 0],
+        [24, 60, 102, 126, 102, 102, 102, 0],
+        [62, 102, 102, 62, 102, 102, 62, 0],
+        [60, 102, 6, 6, 6, 102, 60, 0],
+        [30, 54, 102, 102, 102, 54, 30, 0],
+        [126, 6, 6, 30, 6, 6, 126, 0],
+        [126, 6, 6, 30, 6, 6, 6, 0],
+        [60, 102, 6, 118, 102, 102, 60, 0],
+        [102, 102, 102, 126, 102, 102, 102, 0],
+        [60, 24, 24, 24, 24, 24, 60, 0],
+        [120, 48, 48, 48, 48, 54, 28, 0],
+        [102, 54, 30, 14, 30, 54, 102, 0],
+        [6, 6, 6, 6, 6, 6, 126, 0],
+        [198, 238, 254, 214, 198, 198, 198, 0],
+        [102, 110, 126, 126, 118, 102, 102, 0],
+        [60, 102, 102, 102, 102, 102, 60, 0],
+        [62, 102, 102, 62, 6, 6, 6, 0],
+        [60, 102, 102, 102, 102, 60, 112, 0],
+        [62, 102, 102, 62, 30, 54, 102, 0],
+        [60, 102, 6, 60, 96, 102, 60, 0],
+        [126, 24, 24, 24, 24, 24, 24, 0],
+        [102, 102, 102, 102, 102, 102, 60, 0],
+        [102, 102, 102, 102, 102, 60, 24, 0],
+        [198, 198, 198, 214, 254, 238, 198, 0],
+        [102, 102, 60, 24, 60, 102, 102, 0],
+        [102, 102, 102, 60, 24, 24, 24, 0],
+        [126, 96, 48, 24, 12, 6, 126, 0],
+        [60, 12, 12, 12, 12, 12, 60, 0],
+        [48, 72, 12, 62, 12, 70, 63, 0],
+        [60, 48, 48, 48, 48, 48, 60, 0],
+        [0, 24, 60, 126, 24, 24, 24, 24],
+        [0, 8, 12, 254, 254, 12, 8, 0],
+        [0, 0, 0, 0, 0, 0, 0, 0],
+        [24, 24, 24, 24, 0, 0, 24, 0],
+        [102, 102, 102, 0, 0, 0, 0, 0],
+        [102, 102, 255, 102, 255, 102, 102, 0],
+        [24, 124, 6, 60, 96, 62, 24, 0],
+        [70, 102, 48, 24, 12, 102, 98, 0],
+        [60, 102, 60, 28, 230, 102, 252, 0],
+        [96, 48, 24, 0, 0, 0, 0, 0],
+        [48, 24, 12, 12, 12, 24, 48, 0],
+        [12, 24, 48, 48, 48, 24, 12, 0],
+        [0, 102, 60, 255, 60, 102, 0, 0],
+        [0, 24, 24, 126, 24, 24, 0, 0],
+        [0, 0, 0, 0, 0, 24, 24, 12],
+        [0, 0, 0, 126, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0, 24, 24, 0],
+        [0, 192, 96, 48, 24, 12, 6, 0],
+        [60, 102, 118, 110, 102, 102, 60, 0],
+        [24, 24, 28, 24, 24, 24, 126, 0],
+        [60, 102, 96, 48, 12, 6, 126, 0],
+        [60, 102, 96, 56, 96, 102, 60, 0],
+        [96, 112, 120, 102, 254, 96, 96, 0],
+        [126, 6, 62, 96, 96, 102, 60, 0],
+        [60, 102, 6, 62, 102, 102, 60, 0],
+        [126, 102, 48, 24, 24, 24, 24, 0],
+        [60, 102, 102, 60, 102, 102, 60, 0],
+        [60, 102, 102, 124, 96, 102, 60, 0],
+        [0, 0, 24, 0, 0, 24, 0, 0],
+        [0, 0, 24, 0, 0, 24, 24, 12],
+        [112, 24, 12, 6, 12, 24, 112, 0],
+        [0, 0, 126, 0, 126, 0, 0, 0],
+        [14, 24, 48, 96, 48, 24, 14, 0],
+        [60, 102, 96, 48, 24, 0, 24, 0],
+        [0, 0, 0, 255, 255, 0, 0, 0],
+        [16, 56, 124, 254, 254, 56, 124, 0],
+        [24, 24, 24, 24, 24, 24, 24, 24],
+        [0, 0, 0, 255, 255, 0, 0, 0],
+        [0, 0, 255, 255, 0, 0, 0, 0],
+        [0, 255, 255, 0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 255, 255, 0, 0],
+        [12, 12, 12, 12, 12, 12, 12, 12],
+        [48, 48, 48, 48, 48, 48, 48, 48],
+        [0, 0, 0, 7, 15, 28, 24, 24],
+        [24, 24, 56, 240, 224, 0, 0, 0],
+        [24, 24, 28, 15, 7, 0, 0, 0],
+        [3, 3, 3, 3, 3, 3, 255, 255],
+        [3, 7, 14, 28, 56, 112, 224, 192],
+        [192, 224, 112, 56, 28, 14, 7, 3],
+        [255, 255, 3, 3, 3, 3, 3, 3],
+        [255, 255, 192, 192, 192, 192, 192, 192],
+        [0, 60, 126, 126, 126, 126, 60, 0],
+        [0, 0, 0, 0, 0, 255, 255, 0],
+        [108, 254, 254, 254, 124, 56, 16, 0],
+        [6, 6, 6, 6, 6, 6, 6, 6],
+        [0, 0, 0, 224, 240, 56, 24, 24],
+        [195, 231, 126, 60, 60, 126, 231, 195],
+        [0, 60, 126, 102, 102, 126, 60, 0],
+        [24, 24, 102, 102, 24, 24, 60, 0],
+        [96, 96, 96, 96, 96, 96, 96, 96],
+        [16, 56, 124, 254, 124, 56, 16, 0],
+        [24, 24, 24, 255, 255, 24, 24, 24],
+        [3, 3, 12, 12, 3, 3, 12, 12],
+        [24, 24, 24, 24, 24, 24, 24, 24],
+        [0, 0, 192, 124, 110, 108, 108, 0],
+        [255, 254, 252, 248, 240, 224, 192, 128],
+        [0, 0, 0, 0, 0, 0, 0, 0],
+        [15, 15, 15, 15, 15, 15, 15, 15],
+        [0, 0, 0, 0, 255, 255, 255, 255],
+        [255, 0, 0, 0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0, 0, 0, 255],
+        [3, 3, 3, 3, 3, 3, 3, 3],
+        [51, 51, 204, 204, 51, 51, 204, 204],
+        [192, 192, 192, 192, 192, 192, 192, 192],
+        [0, 0, 0, 0, 51, 51, 204, 204],
+        [255, 127, 63, 31, 15, 7, 3, 1],
+        [192, 192, 192, 192, 192, 192, 192, 192],
+        [24, 24, 24, 248, 248, 24, 24, 24],
+        [0, 0, 0, 0, 240, 240, 240, 240],
+        [24, 24, 24, 248, 248, 0, 0, 0],
+        [0, 0, 0, 31, 31, 24, 24, 24],
+        [0, 0, 0, 0, 0, 0, 255, 255],
+        [0, 0, 0, 248, 248, 24, 24, 24],
+        [24, 24, 24, 255, 255, 0, 0, 0],
+        [0, 0, 0, 255, 255, 24, 24, 24],
+        [24, 24, 24, 31, 31, 24, 24, 24],
+        [3, 3, 3, 3, 3, 3, 3, 3],
+        [7, 7, 7, 7, 7, 7, 7, 7],
+        [224, 224, 224, 224, 224, 224, 224, 224],
+        [255, 255, 0, 0, 0, 0, 0, 0],
+        [255, 255, 255, 0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0, 255, 255, 255],
+        [192, 192, 192, 192, 192, 192, 255, 255],
+        [0, 0, 0, 0, 15, 15, 15, 15],
+        [240, 240, 240, 240, 0, 0, 0, 0],
+        [24, 24, 24, 31, 31, 0, 0, 0],
+        [15, 15, 15, 15, 0, 0, 0, 0],
+        [15, 15, 15, 15, 240, 240, 240, 240],
+        [195, 153, 137, 137, 249, 153, 195, 255],
+        [231, 195, 153, 129, 153, 153, 153, 255],
+        [193, 153, 153, 193, 153, 153, 193, 255],
+        [195, 153, 249, 249, 249, 153, 195, 255],
+        [225, 201, 153, 153, 153, 201, 225, 255],
+        [129, 249, 249, 225, 249, 249, 129, 255],
+        [129, 249, 249, 225, 249, 249, 249, 255],
+        [195, 153, 249, 137, 153, 153, 195, 255],
+        [153, 153, 153, 129, 153, 153, 153, 255],
+        [195, 231, 231, 231, 231, 231, 195, 255],
+        [135, 207, 207, 207, 207, 201, 227, 255],
+        [153, 201, 225, 241, 225, 201, 153, 255],
+        [249, 249, 249, 249, 249, 249, 129, 255],
+        [57, 17, 1, 41, 57, 57, 57, 255],
+        [153, 145, 129, 129, 137, 153, 153, 255],
+        [195, 153, 153, 153, 153, 153, 195, 255],
+        [193, 153, 153, 193, 249, 249, 249, 255],
+        [195, 153, 153, 153, 153, 195, 143, 255],
+        [193, 153, 153, 193, 225, 201, 153, 255],
+        [195, 153, 249, 195, 159, 153, 195, 255],
+        [129, 231, 231, 231, 231, 231, 231, 255],
+        [153, 153, 153, 153, 153, 153, 195, 255],
+        [153, 153, 153, 153, 153, 195, 231, 255],
+        [57, 57, 57, 41, 1, 17, 57, 255],
+        [153, 153, 195, 231, 195, 153, 153, 255],
+        [153, 153, 153, 195, 231, 231, 231, 255],
+        [129, 159, 207, 231, 243, 249, 129, 255],
+        [195, 243, 243, 243, 243, 243, 195, 255],
+        [207, 183, 243, 193, 243, 185, 192, 255],
+        [195, 207, 207, 207, 207, 207, 195, 255],
+        [255, 231, 195, 129, 231, 231, 231, 231],
+        [255, 247, 243, 1, 1, 243, 247, 255],
+        [255, 255, 255, 255, 255, 255, 255, 255],
+        [231, 231, 231, 231, 255, 255, 231, 255],
+        [153, 153, 153, 255, 255, 255, 255, 255],
+        [153, 153, 0, 153, 0, 153, 153, 255],
+        [231, 131, 249, 195, 159, 193, 231, 255],
+        [185, 153, 207, 231, 243, 153, 157, 255],
+        [195, 153, 195, 227, 25, 153, 3, 255],
+        [159, 207, 231, 255, 255, 255, 255, 255],
+        [207, 231, 243, 243, 243, 231, 207, 255],
+        [243, 231, 207, 207, 207, 231, 243, 255],
+        [255, 153, 195, 0, 195, 153, 255, 255],
+        [255, 231, 231, 129, 231, 231, 255, 255],
+        [255, 255, 255, 255, 255, 231, 231, 243],
+        [255, 255, 255, 129, 255, 255, 255, 255],
+        [255, 255, 255, 255, 255, 231, 231, 255],
+        [255, 63, 159, 207, 231, 243, 249, 255],
+        [195, 153, 137, 145, 153, 153, 195, 255],
+        [231, 231, 227, 231, 231, 231, 129, 255],
+        [195, 153, 159, 207, 243, 249, 129, 255],
+        [195, 153, 159, 199, 159, 153, 195, 255],
+        [159, 143, 135, 153, 1, 159, 159, 255],
+        [129, 249, 193, 159, 159, 153, 195, 255],
+        [195, 153, 249, 193, 153, 153, 195, 255],
+        [129, 153, 207, 231, 231, 231, 231, 255],
+        [195, 153, 153, 195, 153, 153, 195, 255],
+        [195, 153, 153, 131, 159, 153, 195, 255],
+        [255, 255, 231, 255, 255, 231, 255, 255],
+        [255, 255, 231, 255, 255, 231, 231, 243],
+        [143, 231, 243, 249, 243, 231, 143, 255],
+        [255, 255, 129, 255, 129, 255, 255, 255],
+        [241, 231, 207, 159, 207, 231, 241, 255],
+        [195, 153, 159, 207, 231, 255, 231, 255],
+        [255, 255, 255, 0, 0, 255, 255, 255],
+        [239, 199, 131, 1, 1, 199, 131, 255],
+        [231, 231, 231, 231, 231, 231, 231, 231],
+        [255, 255, 255, 0, 0, 255, 255, 255],
+        [255, 255, 0, 0, 255, 255, 255, 255],
+        [255, 0, 0, 255, 255, 255, 255, 255],
+        [255, 255, 255, 255, 0, 0, 255, 255],
+        [243, 243, 243, 243, 243, 243, 243, 243],
+        [207, 207, 207, 207, 207, 207, 207, 207],
+        [255, 255, 255, 248, 240, 227, 231, 231],
+        [231, 231, 199, 15, 31, 255, 255, 255],
+        [231, 231, 227, 240, 248, 255, 255, 255],
+        [252, 252, 252, 252, 252, 252, 0, 0],
+        [252, 248, 241, 227, 199, 143, 31, 63],
+        [63, 31, 143, 199, 227, 241, 248, 252],
+        [0, 0, 252, 252, 252, 252, 252, 252],
+        [0, 0, 63, 63, 63, 63, 63, 63],
+        [255, 195, 129, 129, 129, 129, 195, 255],
+        [255, 255, 255, 255, 255, 0, 0, 255],
+        [147, 1, 1, 1, 131, 199, 239, 255],
+        [249, 249, 249, 249, 249, 249, 249, 249],
+        [255, 255, 255, 31, 15, 199, 231, 231],
+        [60, 24, 129, 195, 195, 129, 24, 60],
+        [255, 195, 129, 153, 153, 129, 195, 255],
+        [231, 231, 153, 153, 231, 231, 195, 255],
+        [159, 159, 159, 159, 159, 159, 159, 159],
+        [239, 199, 131, 1, 131, 199, 239, 255],
+        [231, 231, 231, 0, 0, 231, 231, 231],
+        [252, 252, 243, 243, 252, 252, 243, 243],
+        [231, 231, 231, 231, 231, 231, 231, 231],
+        [255, 255, 63, 131, 145, 147, 147, 255],
+        [0, 1, 3, 7, 15, 31, 63, 127],
+        [255, 255, 255, 255, 255, 255, 255, 255],
+        [240, 240, 240, 240, 240, 240, 240, 240],
+        [255, 255, 255, 255, 0, 0, 0, 0],
+        [0, 255, 255, 255, 255, 255, 255, 255],
+        [255, 255, 255, 255, 255, 255, 255, 0],
+        [252, 252, 252, 252, 252, 252, 252, 252],
+        [204, 204, 51, 51, 204, 204, 51, 51],
+        [63, 63, 63, 63, 63, 63, 63, 63],
+        [255, 255, 255, 255, 204, 204, 51, 51],
+        [0, 128, 192, 224, 240, 248, 252, 254],
+        [63, 63, 63, 63, 63, 63, 63, 63],
+        [231, 231, 231, 7, 7, 231, 231, 231],
+        [255, 255, 255, 255, 15, 15, 15, 15],
+        [231, 231, 231, 7, 7, 255, 255, 255],
+        [255, 255, 255, 224, 224, 231, 231, 231],
+        [255, 255, 255, 255, 255, 255, 0, 0],
+        [255, 255, 255, 7, 7, 231, 231, 231],
+        [231, 231, 231, 0, 0, 255, 255, 255],
+        [255, 255, 255, 0, 0, 231, 231, 231],
+        [231, 231, 231, 224, 224, 231, 231, 231],
+        [252, 252, 252, 252, 252, 252, 252, 252],
+        [248, 248, 248, 248, 248, 248, 248, 248],
+        [31, 31, 31, 31, 31, 31, 31, 31],
+        [0, 0, 255, 255, 255, 255, 255, 255],
+        [0, 0, 0, 255, 255, 255, 255, 255],
+        [255, 255, 255, 255, 255, 0, 0, 0],
+        [63, 63, 63, 63, 63, 63, 0, 0],
+        [255, 255, 255, 255, 240, 240, 240, 240],
+        [15, 15, 15, 15, 255, 255, 255, 255],
+        [231, 231, 231, 224, 224, 255, 255, 255],
+        [240, 240, 240, 240, 255, 255, 255, 255],
+        [240, 240, 240, 240, 15, 15, 15, 15],
+        [60, 102, 118, 118, 6, 70, 60, 0],
+        [0, 0, 60, 96, 124, 102, 124, 0],
+        [0, 6, 6, 62, 102, 102, 62, 0],
+        [0, 0, 60, 6, 6, 6, 60, 0],
+        [0, 96, 96, 124, 102, 102, 124, 0],
+        [0, 0, 60, 102, 126, 6, 60, 0],
+        [0, 112, 24, 124, 24, 24, 24, 0],
+        [0, 0, 124, 102, 102, 124, 96, 62],
+        [0, 6, 6, 62, 102, 102, 102, 0],
+        [0, 24, 0, 28, 24, 24, 60, 0],
+        [0, 96, 0, 96, 96, 96, 96, 60],
+        [0, 6, 6, 54, 30, 54, 102, 0],
+        [0, 28, 24, 24, 24, 24, 60, 0],
+        [0, 0, 102, 254, 254, 214, 198, 0],
+        [0, 0, 62, 102, 102, 102, 102, 0],
+        [0, 0, 60, 102, 102, 102, 60, 0],
+        [0, 0, 62, 102, 102, 62, 6, 6],
+        [0, 0, 124, 102, 102, 124, 96, 96],
+        [0, 0, 62, 102, 6, 6, 6, 0],
+        [0, 0, 124, 6, 60, 96, 62, 0],
+        [0, 24, 126, 24, 24, 24, 112, 0],
+        [0, 0, 102, 102, 102, 102, 124, 0],
+        [0, 0, 102, 102, 102, 60, 24, 0],
+        [0, 0, 198, 214, 254, 124, 108, 0],
+        [0, 0, 102, 60, 24, 60, 102, 0],
+        [0, 0, 102, 102, 102, 124, 48, 30],
+        [0, 0, 126, 48, 24, 12, 126, 0],
+        [60, 12, 12, 12, 12, 12, 60, 0],
+        [48, 72, 12, 62, 12, 70, 63, 0],
+        [60, 48, 48, 48, 48, 48, 60, 0],
+        [0, 24, 60, 126, 24, 24, 24, 24],
+        [0, 8, 12, 254, 254, 12, 8, 0],
+        [0, 0, 0, 0, 0, 0, 0, 0],
+        [24, 24, 24, 24, 0, 0, 24, 0],
+        [102, 102, 102, 0, 0, 0, 0, 0],
+        [102, 102, 255, 102, 255, 102, 102, 0],
+        [24, 124, 6, 60, 96, 62, 24, 0],
+        [70, 102, 48, 24, 12, 102, 98, 0],
+        [60, 102, 60, 28, 230, 102, 252, 0],
+        [96, 48, 24, 0, 0, 0, 0, 0],
+        [48, 24, 12, 12, 12, 24, 48, 0],
+        [12, 24, 48, 48, 48, 24, 12, 0],
+        [0, 102, 60, 255, 60, 102, 0, 0],
+        [0, 24, 24, 126, 24, 24, 0, 0],
+        [0, 0, 0, 0, 0, 24, 24, 12],
+        [0, 0, 0, 126, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0, 24, 24, 0],
+        [0, 192, 96, 48, 24, 12, 6, 0],
+        [60, 102, 118, 110, 102, 102, 60, 0],
+        [24, 24, 28, 24, 24, 24, 126, 0],
+        [60, 102, 96, 48, 12, 6, 126, 0],
+        [60, 102, 96, 56, 96, 102, 60, 0],
+        [96, 112, 120, 102, 254, 96, 96, 0],
+        [126, 6, 62, 96, 96, 102, 60, 0],
+        [60, 102, 6, 62, 102, 102, 60, 0],
+        [126, 102, 48, 24, 24, 24, 24, 0],
+        [60, 102, 102, 60, 102, 102, 60, 0],
+        [60, 102, 102, 124, 96, 102, 60, 0],
+        [0, 0, 24, 0, 0, 24, 0, 0],
+        [0, 0, 24, 0, 0, 24, 24, 12],
+        [112, 24, 12, 6, 12, 24, 112, 0],
+        [0, 0, 126, 0, 126, 0, 0, 0],
+        [14, 24, 48, 96, 48, 24, 14, 0],
+        [60, 102, 96, 48, 24, 0, 24, 0],
+        [0, 0, 0, 255, 255, 0, 0, 0],
+        [24, 60, 102, 126, 102, 102, 102, 0],
+        [62, 102, 102, 62, 102, 102, 62, 0],
+        [60, 102, 6, 6, 6, 102, 60, 0],
+        [30, 54, 102, 102, 102, 54, 30, 0],
+        [126, 6, 6, 30, 6, 6, 126, 0],
+        [126, 6, 6, 30, 6, 6, 6, 0],
+        [60, 102, 6, 118, 102, 102, 60, 0],
+        [102, 102, 102, 126, 102, 102, 102, 0],
+        [60, 24, 24, 24, 24, 24, 60, 0],
+        [120, 48, 48, 48, 48, 54, 28, 0],
+        [102, 54, 30, 14, 30, 54, 102, 0],
+        [6, 6, 6, 6, 6, 6, 126, 0],
+        [198, 238, 254, 214, 198, 198, 198, 0],
+        [102, 110, 126, 126, 118, 102, 102, 0],
+        [60, 102, 102, 102, 102, 102, 60, 0],
+        [62, 102, 102, 62, 6, 6, 6, 0],
+        [60, 102, 102, 102, 102, 60, 112, 0],
+        [62, 102, 102, 62, 30, 54, 102, 0],
+        [60, 102, 6, 60, 96, 102, 60, 0],
+        [126, 24, 24, 24, 24, 24, 24, 0],
+        [102, 102, 102, 102, 102, 102, 60, 0],
+        [102, 102, 102, 102, 102, 60, 24, 0],
+        [198, 198, 198, 214, 254, 238, 198, 0],
+        [102, 102, 60, 24, 60, 102, 102, 0],
+        [102, 102, 102, 60, 24, 24, 24, 0],
+        [126, 96, 48, 24, 12, 6, 126, 0],
+        [24, 24, 24, 255, 255, 24, 24, 24],
+        [3, 3, 12, 12, 3, 3, 12, 12],
+        [24, 24, 24, 24, 24, 24, 24, 24],
+        [204, 204, 51, 51, 204, 204, 51, 51],
+        [204, 153, 51, 102, 204, 153, 51, 102],
+        [0, 0, 0, 0, 0, 0, 0, 0],
+        [15, 15, 15, 15, 15, 15, 15, 15],
+        [0, 0, 0, 0, 255, 255, 255, 255],
+        [255, 0, 0, 0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0, 0, 0, 255],
+        [3, 3, 3, 3, 3, 3, 3, 3],
+        [51, 51, 204, 204, 51, 51, 204, 204],
+        [192, 192, 192, 192, 192, 192, 192, 192],
+        [0, 0, 0, 0, 51, 51, 204, 204],
+        [51, 153, 204, 102, 51, 153, 204, 102],
+        [192, 192, 192, 192, 192, 192, 192, 192],
+        [24, 24, 24, 248, 248, 24, 24, 24],
+        [0, 0, 0, 0, 240, 240, 240, 240],
+        [24, 24, 24, 248, 248, 0, 0, 0],
+        [0, 0, 0, 31, 31, 24, 24, 24],
+        [0, 0, 0, 0, 0, 0, 255, 255],
+        [0, 0, 0, 248, 248, 24, 24, 24],
+        [24, 24, 24, 255, 255, 0, 0, 0],
+        [0, 0, 0, 255, 255, 24, 24, 24],
+        [24, 24, 24, 31, 31, 24, 24, 24],
+        [3, 3, 3, 3, 3, 3, 3, 3],
+        [7, 7, 7, 7, 7, 7, 7, 7],
+        [224, 224, 224, 224, 224, 224, 224, 224],
+        [255, 255, 0, 0, 0, 0, 0, 0],
+        [255, 255, 255, 0, 0, 0, 0, 0],
+        [0, 0, 0, 0, 0, 255, 255, 255],
+        [128, 192, 96, 54, 30, 14, 6, 0],
+        [0, 0, 0, 0, 15, 15, 15, 15],
+        [240, 240, 240, 240, 0, 0, 0, 0],
+        [24, 24, 24, 31, 31, 0, 0, 0],
+        [15, 15, 15, 15, 0, 0, 0, 0],
+        [15, 15, 15, 15, 240, 240, 240, 240],
+        [195, 153, 137, 137, 249, 153, 195, 255],
+        [255, 255, 195, 159, 131, 153, 131, 255],
+        [255, 249, 249, 193, 153, 153, 193, 255],
+        [255, 255, 195, 249, 249, 249, 195, 255],
+        [255, 159, 159, 131, 153, 153, 131, 255],
+        [255, 255, 195, 153, 129, 249, 195, 255],
+        [255, 143, 231, 131, 231, 231, 231, 255],
+        [255, 255, 131, 153, 153, 131, 159, 193],
+        [255, 249, 249, 193, 153, 153, 153, 255],
+        [255, 231, 255, 227, 231, 231, 195, 255],
+        [255, 159, 255, 159, 159, 159, 159, 195],
+        [255, 249, 249, 201, 225, 201, 153, 255],
+        [255, 227, 231, 231, 231, 231, 195, 255],
+        [255, 255, 153, 1, 1, 41, 57, 255],
+        [255, 255, 193, 153, 153, 153, 153, 255],
+        [255, 255, 195, 153, 153, 153, 195, 255],
+        [255, 255, 193, 153, 153, 193, 249, 249],
+        [255, 255, 131, 153, 153, 131, 159, 159],
+        [255, 255, 193, 153, 249, 249, 249, 255],
+        [255, 255, 131, 249, 195, 159, 193, 255],
+        [255, 231, 129, 231, 231, 231, 143, 255],
+        [255, 255, 153, 153, 153, 153, 131, 255],
+        [255, 255, 153, 153, 153, 195, 231, 255],
+        [255, 255, 57, 41, 1, 131, 147, 255],
+        [255, 255, 153, 195, 231, 195, 153, 255],
+        [255, 255, 153, 153, 153, 131, 207, 225],
+        [255, 255, 129, 207, 231, 243, 129, 255],
+        [195, 243, 243, 243, 243, 243, 195, 255],
+        [207, 183, 243, 193, 243, 185, 192, 255],
+        [195, 207, 207, 207, 207, 207, 195, 255],
+        [255, 231, 195, 129, 231, 231, 231, 231],
+        [255, 247, 243, 1, 1, 243, 247, 255],
+        [255, 255, 255, 255, 255, 255, 255, 255],
+        [231, 231, 231, 231, 255, 255, 231, 255],
+        [153, 153, 153, 255, 255, 255, 255, 255],
+        [153, 153, 0, 153, 0, 153, 153, 255],
+        [231, 131, 249, 195, 159, 193, 231, 255],
+        [185, 153, 207, 231, 243, 153, 157, 255],
+        [195, 153, 195, 227, 25, 153, 3, 255],
+        [159, 207, 231, 255, 255, 255, 255, 255],
+        [207, 231, 243, 243, 243, 231, 207, 255],
+        [243, 231, 207, 207, 207, 231, 243, 255],
+        [255, 153, 195, 0, 195, 153, 255, 255],
+        [255, 231, 231, 129, 231, 231, 255, 255],
+        [255, 255, 255, 255, 255, 231, 231, 243],
+        [255, 255, 255, 129, 255, 255, 255, 255],
+        [255, 255, 255, 255, 255, 231, 231, 255],
+        [255, 63, 159, 207, 231, 243, 249, 255],
+        [195, 153, 137, 145, 153, 153, 195, 255],
+        [231, 231, 227, 231, 231, 231, 129, 255],
+        [195, 153, 159, 207, 243, 249, 129, 255],
+        [195, 153, 159, 199, 159, 153, 195, 255],
+        [159, 143, 135, 153, 1, 159, 159, 255],
+        [129, 249, 193, 159, 159, 153, 195, 255],
+        [195, 153, 249, 193, 153, 153, 195, 255],
+        [129, 153, 207, 231, 231, 231, 231, 255],
+        [195, 153, 153, 195, 153, 153, 195, 255],
+        [195, 153, 153, 131, 159, 153, 195, 255],
+        [255, 255, 231, 255, 255, 231, 255, 255],
+        [255, 255, 231, 255, 255, 231, 231, 243],
+        [143, 231, 243, 249, 243, 231, 143, 255],
+        [255, 255, 129, 255, 129, 255, 255, 255],
+        [241, 231, 207, 159, 207, 231, 241, 255],
+        [195, 153, 159, 207, 231, 255, 231, 255],
+        [255, 255, 255, 0, 0, 255, 255, 255],
+        [231, 195, 153, 129, 153, 153, 153, 255],
+        [193, 153, 153, 193, 153, 153, 193, 255],
+        [195, 153, 249, 249, 249, 153, 195, 255],
+        [225, 201, 153, 153, 153, 201, 225, 255],
+        [129, 249, 249, 225, 249, 249, 129, 255],
+        [129, 249, 249, 225, 249, 249, 249, 255],
+        [195, 153, 249, 137, 153, 153, 195, 255],
+        [153, 153, 153, 129, 153, 153, 153, 255],
+        [195, 231, 231, 231, 231, 231, 195, 255],
+        [135, 207, 207, 207, 207, 201, 227, 255],
+        [153, 201, 225, 241, 225, 201, 153, 255],
+        [249, 249, 249, 249, 249, 249, 129, 255],
+        [57, 17, 1, 41, 57, 57, 57, 255],
+        [153, 145, 129, 129, 137, 153, 153, 255],
+        [195, 153, 153, 153, 153, 153, 195, 255],
+        [193, 153, 153, 193, 249, 249, 249, 255],
+        [195, 153, 153, 153, 153, 195, 143, 255],
+        [193, 153, 153, 193, 225, 201, 153, 255],
+        [195, 153, 249, 195, 159, 153, 195, 255],
+        [129, 231, 231, 231, 231, 231, 231, 255],
+        [153, 153, 153, 153, 153, 153, 195, 255],
+        [153, 153, 153, 153, 153, 195, 231, 255],
+        [57, 57, 57, 41, 1, 17, 57, 255],
+        [153, 153, 195, 231, 195, 153, 153, 255],
+        [153, 153, 153, 195, 231, 231, 231, 255],
+        [129, 159, 207, 231, 243, 249, 129, 255],
+        [231, 231, 231, 0, 0, 231, 231, 231],
+        [252, 252, 243, 243, 252, 252, 243, 243],
+        [231, 231, 231, 231, 231, 231, 231, 231],
+        [51, 51, 204, 204, 51, 51, 204, 204],
+        [51, 102, 204, 153, 51, 102, 204, 153],
+        [255, 255, 255, 255, 255, 255, 255, 255],
+        [240, 240, 240, 240, 240, 240, 240, 240],
+        [255, 255, 255, 255, 0, 0, 0, 0],
+        [0, 255, 255, 255, 255, 255, 255, 255],
+        [255, 255, 255, 255, 255, 255, 255, 0],
+        [252, 252, 252, 252, 252, 252, 252, 252],
+        [204, 204, 51, 51, 204, 204, 51, 51],
+        [63, 63, 63, 63, 63, 63, 63, 63],
+        [255, 255, 255, 255, 204, 204, 51, 51],
+        [204, 102, 51, 153, 204, 102, 51, 153],
+        [63, 63, 63, 63, 63, 63, 63, 63],
+        [231, 231, 231, 7, 7, 231, 231, 231],
+        [255, 255, 255, 255, 15, 15, 15, 15],
+        [231, 231, 231, 7, 7, 255, 255, 255],
+        [255, 255, 255, 224, 224, 231, 231, 231],
+        [255, 255, 255, 255, 255, 255, 0, 0],
+        [255, 255, 255, 7, 7, 231, 231, 231],
+        [231, 231, 231, 0, 0, 255, 255, 255],
+        [255, 255, 255, 0, 0, 231, 231, 231],
+        [231, 231, 231, 224, 224, 231, 231, 231],
+        [252, 252, 252, 252, 252, 252, 252, 252],
+        [248, 248, 248, 248, 248, 248, 248, 248],
+        [31, 31, 31, 31, 31, 31, 31, 31],
+        [0, 0, 255, 255, 255, 255, 255, 255],
+        [0, 0, 0, 255, 255, 255, 255, 255],
+        [255, 255, 255, 255, 255, 0, 0, 0],
+        [127, 63, 159, 201, 225, 241, 249, 255],
+        [255, 255, 255, 255, 240, 240, 240, 240],
+        [15, 15, 15, 15, 255, 255, 255, 255],
+        [231, 231, 231, 224, 224, 255, 255, 255],
+        [240, 240, 240, 240, 255, 255, 255, 255],
+        [240, 240, 240, 240, 15, 15, 15, 15],
+    ],
+}
